requirements.txt
streamlit
ultralytics
opencv-python-headless
pillow
numpy<2.0
